MSAPI-89 | Split OS specific details to classes
