+++ SignNow.Net.Test/SignNow.Net.Test.csproj
@@ -36,6 +36,7 @@
     <Folder Include="UnitTests\Context\" />
     <Folder Include="UnitTests\Exceptions\" />
     <Folder Include="UnitTests\Helpers\" />
+    <Folder Include="UnitTests\Infrastructure\" />
   </ItemGroup>
 
 </Project>
