+++ SignNow.Net/_Internal/Helpers/UserAgentSdkHeaders.cs
@@ -42,26 +42,7 @@ namespace SignNow.Net.Internal.Helpers
         /// <returns></returns>
         public static string OsDetails()
         {
-            var os = "Unknown";
-            var platform = "unknown";
-            var arch = "X64";
-#if NET45
-            os = Environment.OSVersion.ToString();
-            platform = Environment.Is64BitOperatingSystem ? "win64" : "win32";
-            arch = typeof(UserAgentSdkHeaders).Assembly.GetName().ProcessorArchitecture.ToString();
-#else
-            if (RuntimeInformation.IsOSPlatform(OSPlatform.Linux))
-                os = GetLinuxVersion(RuntimeInformation.OSDescription);
-
-            if (RuntimeInformation.IsOSPlatform(OSPlatform.OSX))
-            {
-                os = GetMacOsVersion(RuntimeInformation.OSDescription);
-            }
-
-            platform = RuntimeInformation.ProcessArchitecture.ToString();
-            arch = RuntimeInformation.OSArchitecture.ToString();
-#endif
-            return $"{os}; {platform}; {arch}";
+            return $"{Infrastructure.RuntimeInfo.OsName()}; {Infrastructure.RuntimeInfo.OsPlatform()}; {Infrastructure.RuntimeInfo.OsArch()}";
         }
 
         /// <summary>
@@ -75,116 +56,6 @@ namespace SignNow.Net.Internal.Helpers
 #else
             return $".NET Core/v{RuntimeInformation.FrameworkDescription.Replace(".NET Core ", String.Empty)}";
 #endif
-        }
-
-        public static string GetLinuxVersion(string kernel)
-        {
-            string version = String.Empty;
-#if NETSTANDARD
-            // Linux 4.4.0 - 43 - Microsoft #1-Microsoft Wed Dec 31 14:42:53 PST 2014
-            // Linux 3.10.0-693.21.1.el7.x86_64 #1 SMP Wed Mar 7 19:03:37 UTC 2018
-            var matched = new Regex(@"^(?<kernel>\w+)+\s+(?<version>\d+.?\d+(?:\S)\S+)").Match(kernel.Trim());
-            version = matched.Groups["version"].Value;
-#endif
-            return $"Linux {version}".Trim();
-        }
-
-        public static string GetMacOsVersion(string kernel)
-        {
-            string version = String.Empty;
-
-#if NETSTANDARD
-            /* Darwin 17.5.0 Darwin Kernel Version 17.5.0: Mon Mar  5 22:24:32 PST 2018; root:xnu-4570.51.1~1/RELEASE_X86_64 */
-            var matched = new Regex(@"^(?<kernel>\w+)+\s+(?<major>\d+).(?<minor>\d+).(?<patch>\d+)?").Match(kernel.Trim());
-            var major = int.Parse(matched.Groups["major"].Value);
-            var minor = int.Parse(matched.Groups["minor"].Value);
-
-            switch (major)
-            {
-                case (int)MacOsVersions.MacOsX101:
-                    version = minor == 4 ? "10.1" : "10.0";
-                    break;
-
-                case (int)MacOsVersions.MacOsX102:
-                    version = "10.2";
-                    break;
-
-                case (int)MacOsVersions.MacOsX103:
-                    version = "10.3";
-                    break;
-
-                case (int)MacOsVersions.MacOsX104:
-                    version = "10.4";
-                    break;
-
-                case (int)MacOsVersions.MacOsX105:
-                    version = "10.5";
-                    break;
-
-                case (int)MacOsVersions.MacOsX106:
-                    version = "10.6";
-                    break;
-
-                case (int)MacOsVersions.MacOsX107:
-                    version = "10.7";
-                    break;
-
-                case (int)MacOsVersions.MacOsX108:
-                    version = "10.8";
-                    break;
-
-                case (int)MacOsVersions.MacOsX109:
-                    version = "10.9";
-                    break;
-
-                case (int)MacOsVersions.MacOsX1010:
-                    version = "10.10";
-                    break;
-
-                case (int)MacOsVersions.MacOsX1011:
-                    version = "10.11";
-                    break;
-
-                case (int)MacOsVersions.MacOsX1012:
-                    version = "10.12";
-                    break;
-
-                case (int)MacOsVersions.MacOsX1013:
-                    version = "10.13";
-                    break;
-
-                case (int)MacOsVersions.MacOsX1014:
-                    version = "10.14";
-                    break;
-
-                case (int)MacOsVersions.MacOsX1015:
-                    version = "10.15";
-                    break;
-
-                default:
-                    break;
-            }
-#endif
-            return $"macOs {version}".Trim();
-        }
-
-        internal enum MacOsVersions
-        {
-            MacOsX101 = 1,
-            MacOsX102 = 6,
-            MacOsX103,
-            MacOsX104,
-            MacOsX105,
-            MacOsX106,
-            MacOsX107,
-            MacOsX108,
-            MacOsX109,
-            MacOsX1010,
-            MacOsX1011,
-            MacOsX1012,
-            MacOsX1013,
-            MacOsX1014,
-            MacOsX1015,
         }
 
         /// <summary>
